Update to 1.17
Updated java and dependencies (FAPI and Mixins)

With some incompatible mods (eg. DankStorage) the client will just crash from the server without logging (or a message) from the client or the server; i tried to iron this bug out but I can't find the right motivation, but it's obviously triggered when the client recives the PolyMap.